test_project/tests/requirements.txt
pytest==8.3.3
requests==2.32.3
PyJWT==2.9.0
